test-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file test-requirements.txt test-requirements.in
#
astroid==1.6.0            # via pylint, pylint-plugin-utils
attrs==17.4.0             # via pytest
certifi==2017.11.5        # via requests
chardet==3.0.4            # via requests
codecov==2.0.13
coverage==4.4.2           # via codecov
editdistance==0.4         # via pylint
idna==2.6                 # via requests
isort==4.2.15             # via pylint
lazy-object-proxy==1.3.1  # via astroid
mccabe==0.6.1             # via pylint
pluggy==0.6.0             # via pytest, tox
py==1.5.2                 # via pytest, tox
pycodestyle==2.3.1
pylint-django==0.7.2
pylint-plugin-utils==0.2.6  # via pylint-django
pylint==1.7
pytest-django==3.1.2
pytest==3.3.2             # via pytest-django
requests==2.18.4          # via codecov
six==1.11.0               # via astroid, pylint, pytest, tox
tox==2.9.1
urllib3==1.22             # via requests
virtualenv==15.1.0        # via tox
wrapt==1.10.11            # via astroid
